refactor: recent activities to smaller files working
